moved mainscreen to presentation module
