+++ app/src/main/java/com/example/btonmarket/MainActivity.kt
@@ -3,14 +3,9 @@ package com.example.btonmarket
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Surface
-import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.btonmarket.feature_btonmarket.presentation.home.HomeScreen
+import com.example.btonmarket.feature_btonmarket.presentation.MainScreen
 import com.example.btonmarket.ui.theme.BtonmarketTheme
 
 class MainActivity : ComponentActivity() {
