+++ app/src/main/java/com/example/btonmarket/feature_btonmarket/presentation/MainScreen.kt
@@ -1,4 +1,4 @@
-package com.example.btonmarket
+package com.example.btonmarket.feature_btonmarket.presentation
 
 import androidx.compose.foundation.layout.RowScope
 import androidx.compose.material.*
